fix(indicators): avoid int overflow in ComputeStochastic %D warm-up guard
The %D warm-up guard computed kPeriod - 1 + dPeriod - 1 in unchecked int
arithmetic. A near-int.MaxValue dPeriod (reachable from GetStochasticOscillator,
which only rejects dPeriod < 1) wrapped the threshold negative, so the guard
never fired and the inner loop indexed %K with a large negative index, throwing
ArgumentOutOfRangeException. Computing the threshold in long keeps the documented
all-null %D behavior for a window that can never fill.

closes #2929

+++ src/Equibles.Yahoo.Repositories/TechnicalIndicatorService.cs
@@ -256,7 +256,10 @@ public static class TechnicalIndicatorService
         var d = new List<decimal?>(count);
         for (var i = 0; i < count; i++)
         {
-            if (i < kPeriod - 1 + dPeriod - 1)
+            // Computed in long: kPeriod + dPeriod can exceed int.MaxValue, and an int
+            // overflow here would wrap the threshold negative so the guard never fires
+            // and the inner loop indexes %K out of range.
+            if (i < (long)kPeriod - 1 + dPeriod - 1)
             {
                 d.Add(null);
                 continue;
